# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pulse Expense Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all groups
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a group by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group with members and balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupBody"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Group deleted
  /groups/{id}/members:
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add a member to a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGroupMemberBody"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /settlements:
    get:
      operationId: listSettlements
      tags: [settlements]
      summary: List settlements, optionally filtered by group
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of settlements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
    post:
      operationId: createSettlement
      tags: [settlements]
      summary: Settle a debt between two users in a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSettlementBody"
      responses:
        "201":
          description: Settlement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        upiId:
          type: ["string", "null"]
        avatarColor:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - avatarColor
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        upiId:
          type: ["string", "null"]
        avatarColor:
          type: string
      required:
        - name
        - email
        - avatarColor
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: string
        createdByUserId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - emoji
        - createdByUserId
        - createdAt
    GroupDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: string
        createdByUserId:
          type: integer
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        totalExpenses:
          type: number
        expenseCount:
          type: integer
      required:
        - id
        - name
        - emoji
        - createdByUserId
        - createdAt
        - members
        - totalExpenses
        - expenseCount
    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: string
        createdByUserId:
          type: integer
        memberUserIds:
          type: array
          items:
            type: integer
      required:
        - name
        - emoji
        - createdByUserId
    UpdateGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: string
    GroupMember:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        userId:
          type: integer
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - userId
        - joinedAt
    AddGroupMemberBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId
...
    Settlement:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        fromUserId:
          type: integer
        fromUserName:
          type: string
        toUserId:
          type: integer
        toUserName:
          type: string
        toUserUpiId:
          type: ["string", "null"]
        amount:
          type: number
        upiLink:
          type: ["string", "null"]
        settledAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - fromUserId
        - fromUserName
        - toUserId
        - toUserName
        - amount
        - settledAt
    CreateSettlementBody:
      type: object
      properties:
        groupId:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        amount:
          type: number
      required:
        - groupId
        - fromUserId
        - toUserId
        - amount
...