# lib/api-zod/src/generated/types/groupDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Expense Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface GroupDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  emoji: string;
  createdByUserId: number;
  createdAt: Date;
  members: User[];
  totalExpenses: number;
  expenseCount: number;
}