# lib/api-zod/src/generated/types/updateGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Expense Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateGroupBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  emoji?: string;
}